player

+++ Assets/Scripts/GameData.cs
@@ -4,7 +4,7 @@
     public static int MazeRows = 25;
     public static int MazeColumns = 25;
     public static int CellSize = 4;
-    public static int HalfCellSize = 2;
+    public static float HalfCellSize = 2f;
     public static int EntrancePoint = 0;
 }
